fixed PR comments #1
