+++ frontend/src/routes/modelCatalog/catalogModelDetailsRoute.ts
