+++ frontend/src/routes/modelCatalog/catalogModelRegisterRoutes.ts
