--- frontend/src/routes/modelRegistry/modelDeploymentsRoute.ts
@@ -1,7 +1,0 @@
-import { modelVersionRoute } from './modelVersionsRoute';
-
-export const modelVersionDeploymentsRoute = (
-  mvId: string,
-  rmId?: string,
-  preferredModelRegistry?: string,
-): string => `${modelVersionRoute(mvId, rmId, preferredModelRegistry)}/deployments`;
