+++ frontend/src/routes/modelRegistry/modelVersionsRoute.ts
@@ -8,3 +8,9 @@ export const modelVersionRoute = (
   rmId?: string,
   preferredModelRegistry?: string,
 ): string => `${modelVersionListRoute(rmId, preferredModelRegistry)}/${mvId}`;
+
+export const modelVersionDeploymentsRoute = (
+  mvId: string,
+  rmId?: string,
+  preferredModelRegistry?: string,
+): string => `${modelVersionRoute(mvId, rmId, preferredModelRegistry)}/deployments`;
